# PERF: cache NameMapping