This is the current benchmark compared to the prev release. 

```
benchmark                            old ns/op     new ns/op     delta
BenchmarkLoadValues/dbr_10-8         434935        372940        -14.25%
BenchmarkLoadValues/dbr_100-8        720677        558852        -22.45%
BenchmarkLoadValues/dbr_1000-8       3206205       2310441       -27.94%
BenchmarkLoadValues/dbr_10000-8      26411855      16150062      -38.85%
BenchmarkLoadValues/dbr_100000-8     283680737     147202536     -48.11%

benchmark                            old allocs     new allocs     delta
BenchmarkLoadValues/dbr_10-8         239            123            -48.54%
BenchmarkLoadValues/dbr_100-8        2132           937            -56.05%
BenchmarkLoadValues/dbr_1000-8       21035          9040           -57.02%
BenchmarkLoadValues/dbr_10000-8      210046         90051          -57.13%
BenchmarkLoadValues/dbr_100000-8     2100056        900061         -57.14%

benchmark                            old bytes     new bytes     delta
BenchmarkLoadValues/dbr_10-8         11662         3874          -66.78%
BenchmarkLoadValues/dbr_100-8        100585        22965         -77.17%
BenchmarkLoadValues/dbr_1000-8       986146        210092        -78.70%
BenchmarkLoadValues/dbr_10000-8      10154700      2394698       -76.42%
BenchmarkLoadValues/dbr_100000-8     101280641     23680625      -76.62%
```